// src/patient/common/models/patient-contact-info.model.ts
export class PatientContactInfo {
    PatientMailing?: boolean;
    PatientEmail?: boolean;
    PatientPrimaryPhone?: boolean;
    PatientHomePhone?: boolean;
    PatientMobilePhone?: boolean;
    PatientWorkPhone?: boolean
    ResponsibleMailing?: boolean;
    ResponsibleEmail?: boolean;
    ResponsiblePrimaryPhone?: boolean;
    ResponsibleHomePhone?: boolean;
    ResponsibleMobilePhone?: boolean;
    ResponsibleWorkPhone?: boolean;
}